refactor: extract each service endpoint type
